Level 29 completed

+++ 3.JavaMultithreading/src/com/javarush/task/task09/lesson07/task02/Solution.java
@@ -16,6 +16,6 @@ public class Solution {
     }
 
     public static Integer getValue(boolean first, boolean second) {
-        return first ? 100 : second ? 200 : null;
+        return first ? (Integer) 100 : second ? (Integer) 200 : null;
     }
 }
